requirements.txt
llama-index>=0.11.0
llama-index-vector-stores-faiss>=0.6.0
llama-index-embeddings-huggingface>=0.7.0
faiss-cpu>=1.8.0
rank-bm25>=0.2.2
FlagEmbedding>=1.2.10
PyMuPDF>=1.24.0
pdfplumber>=0.11.0
pydantic>=2.7.0
pytest>=8.0.0
streamlit>=1.35.0
ragas>=0.2.0
